Final Score working
Final score data passed to endScreen.html page by storing it in the url. the new href that the $scope.endGame function sends to includes $scope.score which can then be decoded on the other side. I also put the anuglar code in another js file as i felt it was getting a bit crowded, and i also made a new controller, 'endController', to deal with the endScreen page.